ai.Summarizer/backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-multipart==0.0.6
httpx==0.25.2
openai==1.6.1

# Document processing
PyPDF2==3.0.1
python-docx==1.1.0
beautifulsoup4==4.12.2
requests==2.31.0
lxml==4.9.3

# Text processing
jieba==0.42.1
langdetect==1.0.9
